# contracts/web-api/openapi.yaml
openapi: 3.0.3
info:
  title: SmartView Web API
  description: SmartView Spring Boot 后端对外暴露的 Web API 契约
  version: 1.0.0
  contact:
    name: SmartView Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /api/profile-analyses:
    post:
      tags:
        - 画像分析
      summary: 触发或获取方向画像分析
      description: >
        幂等接口：用户选择面试方向后调用。
        已有同方向成功分析时直接返回；已有进行中任务时返回任务状态；
        分析失败时新建 PROFILE_ANALYZE 任务重新分析。
        创建任务前后端会校验简历向量已成功入库。
      operationId: ensureProfileAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartProfileAnalysisRequest'
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileAnalysisStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/profile-analyses/{profileId}:
    get:
      tags:
        - 画像分析
      summary: 查询画像分析状态
      description: 前端轮询画像分析状态，最多等待 60 秒；成功后才允许开始面试。
      operationId: getProfileAnalysisStatus
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
        - name: roleDirection
          in: query
          required: true
          schema:
            type: string
            enum: [JAVA_BACKEND, AGENT_DEVELOPMENT]
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileAnalysisStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/profile-analyses/{profileId}/retry:
    post:
      tags:
        - 画像分析
      summary: 重试方向画像分析
      description: 画像分析失败时允许用户重试，重试期间不允许开始面试。
      operationId: retryProfileAnalysis
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
        - name: roleDirection
          in: query
          required: true
          schema:
            type: string
            enum: [JAVA_BACKEND, AGENT_DEVELOPMENT]
      responses:
        '200':
          description: 重试任务已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileAnalysisStatusResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证，格式为 Bearer <token>
  schemas:
    ApiResponse:
      type: object
      required:
        - code
        - message
        - traceId
        - timestamp
      properties:
        code:
          type: string
          description: 业务响应码
          example: "SUCCESS"
        message:
          type: string
          description: 响应消息
          example: "操作成功"
        traceId:
          type: string
          format: uuid
          description: 链路追踪 ID
        timestamp:
          type: string
          format: date-time
          description: 响应时间
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - data
        - traceId
        - timestamp
      properties:
        code:
          type: string
          description: 错误码
        message:
          type: string
          description: 错误消息
        data:
          type: object
          nullable: true
          description: 错误响应数据，默认为空
        traceId:
          type: string
          format: uuid
          description: 链路追踪 ID
        timestamp:
          type: string
          format: date-time
          description: 错误发生时间
        details:
          type: object
          description: 错误详细信息
...
    ProfileAnalysisStatusResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ProfileAnalysisStatus'
    ProfileAnalysisStatus:
      type: object
      required:
        - profileId
        - profileVersion
        - roleDirection
        - status
      properties:
        profileAnalysisId:
          type: string
          nullable: true
          description: 画像分析结果 ID，分析成功时返回
        profileId:
          type: string
          description: 简历画像 ID
        profileVersion:
          type: integer
          description: 当前简历画像版本号
        roleDirection:
          type: string
          enum: [JAVA_BACKEND, AGENT_DEVELOPMENT]
          description: 面试方向
        taskId:
          type: string
          nullable: true
          description: 画像分析任务 ID
        status:
          type: string
          enum: [PENDING, PROCESSING, SUCCESS, FAILED, RETRYING]
          description: 画像分析任务状态
        retryCount:
          type: integer
          description: 已使用的重试次数
        errorMessage:
          type: string
          nullable: true
          description: 最近一次失败原因
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: 状态更新时间
    StartProfileAnalysisRequest:
      type: object
      required:
        - profileId
        - roleDirection
      properties:
        profileId:
          type: string
          description: 简历画像 ID
        roleDirection:
          type: string
          enum: [JAVA_BACKEND, AGENT_DEVELOPMENT]
          description: 面试方向
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 资源状态冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: 业务逻辑校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...